+++ lib/utils/validation.dart
@@ -13,9 +13,9 @@ class Validation {
   static String? qty(value, minQty, maxQty) {
     if (isEmpty(value)) {
       return 'This field is required'.tr;
-    } else if (minQty != null && int.parse(value) < minQty) {
+    } else if (minQty != null && double.parse(value) < minQty) {
       return '${'Quantity must be greater than or equal to'.tr} $minQty';
-    } else if (maxQty != null && int.parse(value) > maxQty) {
+    } else if (maxQty != null && double.parse(value) > maxQty) {
       return '${'Quantity must be less than or equal to'.tr} $maxQty';
     }
     return null;
